More utilities modifications
