+++ src/applications/facility-locator/components/search-results-items/common/LocationDirectionsLink.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { buildAddressArray } from '../../../utils/facilityAddress';
+import { buildAddressArray } from 'platform/utilities/facilities-and-mapbox';
 
 function LocationDirectionsLink({ location, from }) {
   let address = buildAddressArray(location);
