+++ src/applications/facility-locator/utils/mapHelpers.js
@@ -1,5 +1,5 @@
 import mbxGeo from '@mapbox/mapbox-sdk/services/geocoding';
-import { mapboxClient } from 'platform/utilities/mapbox';
+import { mapboxClient } from 'platform/utilities/facilities-and-mapbox';
 
 import { BOUNDING_RADIUS } from '../constants';
 
@@ -90,9 +90,6 @@ export const reverseGeocodeBox = (bounds, types = 'address,postcode') => {
   const { lon, lat } = getBoxCenter(bounds);
   return reverseGeocode(lon, lat, types.split(','));
 };
-
-export const staticMapURL = (lat, long, mapboxToken) =>
-  `https://api.mapbox.com/styles/v1/mapbox/streets-v11/static/pin-l+d83933(${long},${lat})/${long},${lat},16/500x300?access_token=${mapboxToken}`;
 
 /**
  * Generates search criteria from lat/long geocoordinates.
