+++ src/applications/static-pages/facilities/tests/NearbyLocations.unit.spec.jsx
@@ -3,8 +3,8 @@ import { Provider } from 'react-redux';
 import { expect } from 'chai';
 import { mount } from 'enzyme';
 import sinon from 'sinon';
-import { mockApiRequest } from 'platform/testing/unit/helpers.js';
-import * as mapboxUtils from 'platform/utilities/mapbox';
+import { mockApiRequest } from 'platform/testing/unit/helpers';
+import * as mapboxUtils from 'platform/utilities/facilities-and-mapbox';
 import NearbyVetCenters from '../vet-center/NearByVetCenters';
 import NearByVALocations from '../vet-center/NearByVALocations';
 import VAFacilityInfoSection from '../vet-center/components/VAFacilityInfoSection';
