+++ src/applications/static-pages/facilities/vet-center/NearByVetCenters.jsx
@@ -1,16 +1,18 @@
 import React, { useState, useEffect } from 'react';
+import { connect, useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
 import { apiRequest } from 'platform/utilities/api';
-import { getFeaturesFromAddress } from 'platform/utilities/mapbox';
-import { connect, useDispatch } from 'react-redux';
+import {
+  calculateBoundingBox,
+  getFeaturesFromAddress,
+} from 'platform/utilities/facilities-and-mapbox';
 import VetCenterInfoSection from './components/VetCenterInfoSection';
 import VetCenterImageSection from './components/VetCenterImageSection';
 import { fetchFacilityStarted, fetchFacilitySuccess } from '../actions';
 import {
-  calculateBoundingBox,
   convertMetersToMiles,
   distancesToNearbyVetCenters,
-} from '../../../facility-locator/utils/facilityDistance';
+} from '../facilityUtilities';
 import buildFacility from './buildFacility';
 
 const NEARBY_VET_CENTER_RADIUS_MILES = 120;
